# docs/swagger.yaml
basePath: /api
definitions:
  models.ApprovalProcess:
    properties:
      approvers:
        items:
          $ref: '#/definitions/models.Approver'
        type: array
      created_at:
        type: string
      document:
        allOf:
        - $ref: '#/definitions/models.Document'
        description: Добавлено поле Document
      document_id:
        type: integer
      id:
        type: integer
      status:
        type: string
    type: object
  models.Approver:
    properties:
      approved_at:
        type: string
      comment:
        type: string
      id:
        type: integer
      process_id:
        type: integer
      status:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Добавлено поле User
      user_id:
        type: integer
    type: object
  models.Document:
    properties:
      completion_date:
        type: string
      contact_person:
        type: string
      created_at:
        type: string
      deadline_date:
        type: string
      file_path:
        type: string
      founder:
        type: string
      founder_inn:
        type: string
      id:
        type: integer
      incoming_number:
        type: string
      kopuk:
        type: integer
      museum_name:
        type: string
      receipt_date:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      esia_id:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для системы согласования документов
  title: Document Approval System API
  version: "1.0"
paths:
  /approvals:
    get:
      consumes:
      - application/json
      description: Получает список согласований для текущего пользователя
      parameters:
      - description: Статус согласования (pending/completed)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ApprovalProcess'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получить согласования пользователя
      tags:
      - approvals
  /approvals/{id}:
    get:
      consumes:
      - application/json
      description: Получает детальную информацию о процессе согласования
      parameters:
      - description: ID процесса согласования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ApprovalProcess'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получить детали согласования
      tags:
      - approvals
  /auth/esia:
    get:
      consumes:
      - application/json
      description: Обработка callback от ЕСИА после авторизации
      parameters:
      - description: Код авторизации
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Callback ЕСИА
      tags:
      - auth
  /documents:
    post:
      consumes:
      - multipart/form-data
      description: Создает новый документ с возможностью прикрепления файла
      parameters:
      - description: Метаданные документа в формате JSON
        in: formData
        name: document
        required: true
        type: string
      - description: Файл документа
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Document'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создать документ
      tags:
      - documents
  /documents/search:
    get:
      consumes:
      - application/json
      description: Поиск документов по параметрам
      parameters:
      - description: Поисковый запрос
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Document'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Поиск документов
      tags:
      - documents
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: Операции с документами
  name: documents
- description: Операции с согласованиями
  name: approvals
- description: Аутентификация через ЕСИА
  name: auth